Extract Base class to its own file
